+++ apk/Changelog.md
@@ -7,6 +7,11 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
+Web Version July 07 2020
+==============
+
+* Display extra languages as (EN-GB for EN or ES-MX for ES) for clips, only the clips Twitch API has support for this extra languages
+
 Apk Version 3.0.217 - July 06 2020
 ==============